+++ src/components/navbar/Logo.tsx
@@ -1,9 +1,12 @@
-import { Link } from "react-router-dom";
+import { Link, useParams } from "react-router-dom";
 import logo from "/src/assets/images/icons/logo_ustr.png";
+import "/src/styles/Logo.css";
 
 const Logo = () => {
+  const { lang } = useParams();
+
   return (
-    <Link className="navbar-brand" to="/home">
+    <Link className="navbar-brand" to={`/${lang}/home`}>
       <div className="logo-wrapper">
         <img src={logo} alt="Logo" height="100" />
       </div>