+++ src/pages/Contact/map.tsx
@@ -1,9 +1,11 @@
 import { Col } from "react-bootstrap";
+import { useTranslation } from "react-i18next";
 
 const Map = () => {
+  const { t } = useTranslation();
   return (
     <Col md={6}>
-      <h2 className="mb-4">Ofis Konumumuz</h2>
+      <h2 className="mb-4">{t("contact.location")}</h2>
       <div style={{ height: "300px", background: "#eee" }}>
         <iframe
           className="w-100 h-100"